tests: Moving Vertex Attribute Fetch OOB
